auth: refactor GenericAuth()
In preparation for a future commit, receive as an argument the identity
of the user associated with the ticket to be created by this function.
Also, introduce a new 'goto done' destructor to make sure that all code
paths go through the same cleanup code.

No functional change is incurred by this commit.

Change-Id: I2f01abc00998baa4f65c2a30143aea93bc1a731d