cmd: Introduce cmd_Tokenize(), cmd_Split(), and cmd_Join()
Add new utility functions to the libcmd library to split and join
strings using a shell-like syntax. This commit adds the new functions
cmd_Tokenize(), cmd_Split(), cmd_FreeSplit(), and cmd_Join(), and
updates the existing cmd_ParseLine() function.

The cmd_Tokenize() function splits a string using a shell-like syntax
into tokens. The length of the tokens are only limited by available
memory. cmd_Tokenize() can be used to populate a linked list or other
collection of tokens.

The cmd_Split() function splits a string into an argv array of pointers
to strings, suitable for exec*()-style functions.

The cmd_Tokenize() and cmd_Split() functions support single and double
quoting in a shell-like manner. For example, the input string:

    "don't worry" be happy

is split into the three tokens:

    don't worry
    be
    happy

cmd_Tokenize() and cmd_Split() support shell-like backslash ('\')
escaping on unix-like platforms.  Backslash escaping is not supported on
Windows by convention to avoid issues with path separators on that
platform.

The cmd_FreeSplit() function is provided to free the resulting
cmd_Split() argv output.

The cmd_Join() function is the complement of cmd_Split(). The cmd_Join()
function will combine the strings in an argv array of strings to create
a properly quoted command line string. The size of the arguments and the
number of arguments is only limited by available memory.

The change updates the existing cmd_ParseLine() funtction to be
implemented with the new cmd_Tokenize() function.  This removes the
token size limit (CMD_TOOBIG error) and provides more complete quoting
support.

Change-Id: I7eb3771aceb81dd91b1a7235c6c03a382815bdd7