ubik: consider AFSCELL env if -cell isn't provided
In addition to remodeling the ugen_* ubik interface, commit 3720f6b6
(Rework the ugen_* interface) introduced a conditional that prevents us
from using the AFSCELL environment variable. Originally, if the -cell
option wasn't provided, we would only use the cell from ThisCell if the
AFSCELL environment variable wasn't defined. After the introduction of
the commit mentioned above, we ignore the environment variable in
question and always use the cell from ThisCell.

To fix this problem, remove the conditional that forces us to always use
the cell from ThisCell when the -cell option isn't provided.