pts: Add option to change the creator of entries
Currently, it is not possible to modify the "creator" field of a user or
group. To make this possible, extend the "pts setfields" command
introducing a new option (-creator) designed to change the field in
question.

Note 1:
In the absence of a PR_GetCapabilities RPC, "pts setfields -creator" is
a no-op for servers that do not support this new option.

Note 2:
The server side of this new functionality will be implemented in a
future commit.