volser: improve readability of RestoreVolume()
Currently, RestoreVolume() removes all existing vnodes that couldn't be
found in the dump file. To do so, this function saves the offset of the
existing vnodes in buffers, clears the entries found in the dump as the
volume is restored, and removes the vnodes associated with the remaining
entries.

To improve the readability of this logic, move the buffers mentioned
above to a struct, rename them, and add comments.

No functional change is incurred by this commit.

Change-Id: If9a60b92479ca5ab825e978c3903e4b8f23d77e8