afs: Pass mtpt to afs_GetVCache if fid is a root
When getting the vcache for the root dir of a given volume, 'parent',
the first argument of afs_GetVCache(), is expected to be NULL. In
preparation for a future commit, instead of NULL, pass the vcache
associated with the mount point of this volume (if available).

No functional change is incurred by this commit.

Change-Id: I8ae18f8bb33feb53de807d0c6a90c283b15fe5bf