afs: Use the fid provided by afs_FSCall()
In preparation for a future commit, let afs_FSCall() specify the target
fid of the RPCs that go through the call returned by this function. For
now, this fid is always equal to the fid associated with the vcache
received as an argument.

No functional change is incurred by this commit.

Change-Id: I25bd5a7292936f8517117eda281dcbdfa686606e