vol: Initialize vnode dv and inode dv consistently
Currently, salvager assumes that the inode data version of a regular
file must be equal to or less than the vnode data version associated
with it. Unfortunately, these fields are initialized differently in
different parts of the code (e,g.: inode_dv = 1 and vnode_dv = 0),
leading salvager to unnecessarily "fix" a healthy volume.

To avoid this false positive, be consistent by always initializing these
fields to the same value.

Change-Id: I251328e88600e859c2536fab51451a98adc604c8