salvager: Don't fix vnode dv of new regular files
Currently, there are two fields recording the Data Version (DV) for
files maintained by the fileserver: the vnode DV, which is stored in the
vnode index and the inode DV, which is stored in the file's metadata
(owner/group/mode bits (OGM) for namei and i_vicepX/di_vicepX for the
inode backend). The vnode DV is updated whenever a file is modified on
the fileserver, and is primarily where we rely on the DV to be stored.
The inode DV is given when the file is created, but is otherwise never
read or written, except by the salvager. If the vnode DV associated with
a given file is smaller than its inode DV, salvager fixes the former and
prints the following message:

  Vnode %d: version < inode version; fixed (old status)

Unfortunately, the fileserver used to initialize the DV for a new file
to 1 in the inode (OGM) DV, and to 0 in the vnode DV, leading salvager
to unnecessarily "fix" a healthy volume. In this case, skip the check
mentioned above to avoid false positives.

Change-Id: Id2256e3647f5ea811d4d11589554c753bb301fde