build: Select param.h in sysname.m4
Currently, the way the file that will be used as param.h is selected
doesn't allow us to add other headers from other architectures, making
cross-compiling difficult. In an effort to facilitate this process,
select the header that will be used as param.h in sysname.m4. By doing
this, we will have more flexibility to choose which headers should be
included (in future commits).

Change-Id: I5c77e9e061b847e770bf25f454fe891b5ead9429