volser: Force full release if going back in time
As explained in commit 8375a7f7, the volserver doesn't prevent restoring
older volume data to a volume. In this scenario, releasing this volume
can also corrupt the read-only copies, since their updateDate will be
greater than the one stored by the read-write volume. Consequently, an
empty dump (without any vnode but with an 'up to date' header) is
delivered and restored on the read-only sites, resulting in a corrupted
volume (header out of sync with vnodes).

Consider the following scenario:

At time T=1:
  Create a volume, populate it, add replicas and release it:
    $ vos create server1 a vol1
    $ fs mkmount /afs/.robotest/mp1 vol1
    $ cd /afs/.robotest/mp1
    $ for i in {1..5} ; do echo "foo bar" >> file$i ; done
    $ vos addsite server1 a vol1
    $ vos addsite server2 a vol1
    $ vos release vol1
  Now, dump vol1:
    $ vos dump vol1 -file /tmp/vol1.dump

At time T=2:
  Change the files created earlier and release the volume:
    $ cd /afs/.robotest/mp1
    $ for i in {1..5} ; do cat /dev/null > file$i ; done
    $ vos release vol1

At time T=3:
  Perform an incremental restore of vol1.dump and release the volume:
    $ vos restore server1 a vol1 -file /tmp/vol1.dump -overwrite i
    $ vos release vol1

Notes:
  - An incremental restore doesn't change the volume's creation date. As
    a result, vos fails to notice that the volume has been replaced and
    a full release is required.
  - Consequently, only the vnodes with a modification date greater than
    the updateDate of the replicas are delivered.
  - Since the read-write volume has been restored to an earlier version,
    no vnode has a modification date greater than the updateDate of the
    replicas.
  - As a result, an empty dump (without any vnode but with an up to date
    header) is delivered and restored on the read-only sites, corrupting
    the read-only copies.

At time T=4:
  Replicas don't have the same content as the read-write volume:
    $ cd /afs/robotest/mp1
    $ cat ./file1
    $ cd /afs/.robotest/mp1
    $ cat ./file1
    foo bar

To avoid this problem, force a full release if the updateDate of the
read-only copy(ies) is greater than the one stored by the read-write
volume.

Change-Id: I627dc5c16014268ab03a9d77f0a6ebd51595bb43