afs: Ensure dirEntry->name has room for entry name
Currently, the afs_dynroot_addDirEnt() function assumes that the caller
has allocated the directory to be large enough to hold the necessary
entry. In order to make sure that memory isn't being accidentally
stomped, panic if the directory in question doesn't have enough space
for the name of the entry received as an argument. That way, if the
caller makes a mistake we don't fail silently.

Change-Id: I9c9b0ee6bc32ae7751005dd9b21a75e878e59fe3