afs: Speed up removal of vcaches from afs_vhashT
In order to remove an entry from the afs_vhashT hash table, we have to:

1. Find the bucket where the entry to be removed should be present.
2. Go through each element of this bucket until we find this entry.
3. Remove it.

Due to step number 2, the time required to remove an entry from a bucket
increases as the linked list associated with this bucket gets longer.

To avoid this problem, build afs_vhashT using the 'afs_q' data structure
(like commit 4fc48af8 did for afs_vhashTV). Doing so, removing a vcache
from a bucket is just a matter of moving pointers (QRemove()).

Change-Id: I5a1a9a090f9aa3d8884e2bb12aca1f1acc3b902d