afs: Flush unlinked vcaches sooner
On some platforms (such as LINUX), unliked vcaches are not destroyed
until the next execution of afs_ShakeLooseVCaches(). Since the hash for
a given vcache is computed from its volume id and vnode number, keeping
unliked vcaches around can be a problem if files are repeatedly deleted
and recreated in the same volume. In this scenario, the same vnode
number will be reused several times, resulting in many vcaches with the
same volume id and vnode number (but different unique ids) in the same
bucket. Consequently, finding a given vcache in the bucket in question
can be time consuming (and cause extra cpu processing), as we would have
to traverse a long linked list.

To avoid this problem, before creating a new vcache, check if we have
unliked vcaches that haven't been flushed yet. If so, flush them all.

Change-Id: Idc74da0c3c0b27bfdf2cd491b003bdd42c889a95