afs: Make FlushReclaimedVcaches() Darwin specific
Currently, the ReclaimedVCList is only fed on Darwin. As a result, the
logic present in afs_FlushReclaimedVcaches() is never used on all other
platforms and should be wrapped in #ifdef AFS_DARWIN_ENV. While here,
also reorganize this function in an effort to make it more readable.

No functional change is incurred by this commit.

Change-Id: I1097a2b07fcb48dec2b31385477c34e974144b51