tests: Verify cmd_ParseLine() output in libcmd tests
Verify the cmd_ParseLine() output is correct in the existing libcmd unit
tests.

Add tests to verify cmd_ParseLine() fails with CMD_TOOMANY when the argv
array to be filled is not large enough, and succeeds the array is just
large enough.

Add a test to verify cmd_ParseLine() fails with CMD_TOOBIG when it
encounters a token that exceeds its internal token buffer size.

Change-Id: Ia9634ad15e9a10d14a95b1a53f40fcb8f15b6ad2