vol: Optionally remove salvaged RW volumes
A common AFS volume management pattern is to implement a central cell
("ReadWrite" or "RW" cell) where volume content is curated and then
distributed to other remote cells ("ReadOnly" or "RO" cells).

In this pattern, the canonical version of a volume's content is NOT the
parent RW that resides on the RO cell, but the RW in the original RW
cell.  Therefore, if a salvage should happen to repair or change the
contents of the parent RW on an RO cell, the volume may no longer match
the original on the RW cell.

In this case it may be preferable to delete the RW volume that was
modified by salvager, and then restore it (redistribute it) from the
central RW cell.

In support of this use case, provide a new option for salvages:
  -on-rw-error <disposition>
where <disposition> may be:
  "remove"  remove (purge) an RW volume if it is changed by salvager
  "keep"    keep changed RW volumes (default; current behavior)

This option is supported by:
 - bos salvage
 - salvager
 - dasalvager
 - salvageserver

If -on-rw-error remove is specified, and a RW volume is changed by
salvager and removed, any other members of the volume group (RO, BK) are
unaffected.

Change-Id: Ia5c1ebb68c78c05e32dbd5c06468261336cf09e5