Ubik: Move transaction states to their own field
Each transaction has a field called 'flags' that represents, among other
things, the current state of the transaction in question. Since this
field can be modified not only by the transaction's thread but also by
the recovery thread (urecovery_AbortAll -> udisk_abort), it must be
accessed under the database lock. Unfortunately, ubik_trans->flags is
modified without any protection in some functions (ubik_AbortTrans,
ubik_EndTrans, and ubik_CheckCache). As a result, if two threads modify
this field at the same time, the flags stored by it may not correctly
represent the real state of the transaction in question. For example, a
transaction that should be aborted can be successfully completed.

Rather than fixing this problem by always accessing this field under the
database lock, move the flags that are used by other threads to their
own field (states). This way, we can solve the problem mentioned above
without increasing the contention of the database lock.

Change-Id: I433c71609bd4e2358cb2a65feaddc8e66359ac02