ubik: Invalidate app cache if trans was aborted
Before proceeding, each transaction checks if the application cache is
in sync with the database (dbase->cachedVersion == dbase->version).
If not, we:

1a. Update the cache (ubik_updatecache_func).
1b. Update the cache version (dbase->cachedVersion = dbase->version).

The steps mentioned above are performed under the cache lock but without
the database lock. As result, the cache can be updated while receiving a
new database from another site (SDISK_SendFile). If we receive a new
database between steps 1a and 1b, our dbase->cachedVersion will not
correctly represent the version of the data stored in our cache. This
problem is momentary, since the transaction in question will be aborted
(SDISK_SendFile -> urecovery_AbortAll) and ubik_AbortTrans() invalidates
the cache. However, until then, all read transactions running during
that interval will access a corrupted cache (out-of-sync).

To fix this problem, only set dbase->cachedVersion to the current
version if the transaction in question wasn't aborted. While here, also
make sure that ubik_CacheUpdate() reads dbase->version under the
database lock, as this field can be changed by the recovery thread.

Change-Id: Ibdf51e20fba7707eac4832770e09072b113818e7