tests: rx/perf wait for server init before client
On slow or loaded systems, the rx/perf test could occasionally fail due
to the rxperf server not being fully initialized before the client
started. This caused test errors, even without changes to the rx code.

 make CHECK TESTS="-v rx/perf"
 ...
 rx/perf

 1..4
 ok 1 - Started rxperf server
 not ok 2 - single threaded client ran successfully
 RPC: threads	30, times	1, write bytes	1048576, read bytes...
 ok 3 - multi threaded client ran succesfully
 ok 4 - Server exited succesfully
 FAILED 2 (exit status 1)

Add a small delay after starting the server to ensure it has time to
initialize.

Note: This failure was observed in an OpenAFS buildbot worker that
included a make tests, and which would occasionally fail when there was
no rx related code changes. The intermittent failure could be duplicated
on a slower virtual test system, but would not fail on a faster system.

Change-Id: Ie11e0d726ce287c45a677f3bb799388121aafc1e