Introduce gklog
Add a new client utility, called gklog, which is effectively the rxgk
replacement for aklog. The current implementation is very minimal, and
does not implement aklog's various features like logging to paths or
multiple cells, using .xlog, etc. This just acquires tokens for a
single cell, the bare minimum for using authenticated rxgk access with
regular tokens.

This commit also adds the bits of rxgk that are necessary for gklog
functionality, namely the client-side bits of the GSSNegotiate loop
(rxgk_NegotiateClientToken), and a couple of other minor pieces.

In order to implement this, this commit also introduces
ugen_ClientInitService, a new variant to the ugen_ClientInit* family
of functions. This lets us call ugen_ClientInit with a specific
service id to connect to. Adjust the internal helpers a bit so we
don't need to specify values for 'deadtime' and 'maxservers', and get
rid of the unused confusing 'serviceid' argument.

Change-Id: I8a58c1c0129c59b2ba9ee4665aa6ef4c8c54ae71